# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Invosit API
  description: |
    Git sidecar for files that shouldn't be in git.
    Invosit lets teams push and pull encrypted files alongside a repo,
    with per-user access control enforced via envelope encryption.
  version: 0.1.0
  contact:
    name: Invosit
servers:
  - url: https://api.invosit.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
...
paths:
...
  /workspaces/{workspaceId}/environments/{environmentId}/files:
    get:
      tags: [Files]
      summary: List files in environment
      description: Returns file metadata the current user has access to. Does not return file content.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/environmentId"
      responses:
        "200":
          description: List of tracked files
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      $ref: "#/components/schemas/FileMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Files]
      summary: Push a file (register metadata + get signed upload URL)
      description: |
        Registers a new version of a file in the environment and returns a
        short-lived signed PUT URL the CLI uses to upload the (encrypted)
        blob directly to storage. The DB row is created before the URL is
        issued — a failed upload leaves an orphan version, not a missing one.

        **MVP status (issue #11):** unencrypted iteration. `content_hash` is
        the sha256 of the plaintext. M4 will add `wrapped_deks` to the request
        and make `content_hash` the ciphertext hash.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/environmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilePushRequest"
      responses:
        "201":
          description: File pushed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilePushResponse"
        "400":
          description: Invalid path, hash, or size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /workspaces/{workspaceId}/environments/{environmentId}/files/{fileId}:
    get:
      tags: [Files]
      summary: Get a signed download URL for a file
      description: |
        Returns a short-lived signed URL (15 min expiry) plus the file's
        metadata so the CLI can download (and, post-M4, decrypt) it.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/environmentId"
        - $ref: "#/components/parameters/fileId"
      responses:
        "200":
          description: Signed URL and metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDownloadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    delete:
      tags: [Files]
      summary: Delete a file from the environment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/workspaceId"
        - $ref: "#/components/parameters/environmentId"
        - $ref: "#/components/parameters/fileId"
      responses:
        "204":
          description: File deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Kratos session token
      description: |
        Bearer token issued by Ory Kratos and obtained via the CLI's Kratos
        API flow (password today; OIDC / passkey / SSO via API flow with
        token-exchange code in the future). The API forwards this header to
        Kratos `/sessions/whoami` to validate the session.
    cookieAuth:
      type: apiKey
      in: cookie
      name: ory_kratos_session
      description: |
        Kratos session cookie (used by browser clients). The API forwards
        the request's `Cookie` header to Kratos `/sessions/whoami`.
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
      example: ws_abc123
    fileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
      example: file_xyz789
    environmentId:
      name: environmentId
      in: path
      required: true
      schema:
        type: string
      example: env_def456
...
  responses:
    Unauthorized:
      description: Missing or invalid Kratos session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: resource not found
        code:
          type: string
          example: NOT_FOUND
...
    FileMeta:
      type: object
      properties:
        id:
          type: string
          example: file_xyz789
        environment_id:
          type: string
          example: env_def456
        path:
          type: string
          description: Relative path as it will appear on disk
          example: config/example.txt
        content_hash:
          type: string
          description: SHA-256 hash of the file content (plaintext in MVP, ciphertext post-M4)
          example: a1b2c3d4e5f6...
        size:
          type: integer
          description: File size in bytes
          example: 4096
        pushed_by:
          type: string
          example: usr_def456
        pushed_at:
          type: string
          format: date-time
    FilePushRequest:
      type: object
      required: [path, content_hash, size]
      properties:
        path:
          type: string
          description: Relative path the file should be written to on pull. Must not be absolute, contain "..", null bytes, or start with a separator.
          example: config/example.txt
        content_hash:
          type: string
          description: SHA-256 hash of the file content (plaintext in MVP, ciphertext post-M4). 64 lowercase hex chars.
          example: a1b2c3d4e5f6...
        size:
          type: integer
          description: File size in bytes
          example: 4096
        message:
          type: string
          description: Optional version message
          example: rotated API token
    FilePushResponse:
      type: object
      description: |
        Returned on successful push. The CLI uploads the (encrypted) blob
        via a PUT to `upload_url` before the URL expires.
      properties:
        id:
          type: string
          example: file_xyz789
        environment_id:
          type: string
          example: env_def456
        path:
          type: string
          example: config/example.txt
        content_hash:
          type: string
          example: a1b2c3d4e5f6...
        size:
          type: integer
          example: 4096
        pushed_by:
          type: string
          example: usr_def456
        pushed_at:
          type: string
          format: date-time
        version_id:
          type: string
          example: ver_9x8y7z
        upload_url:
          type: string
          description: Short-lived signed PUT URL (15 min). Upload the blob here.
          example: https://r2.invosit.dev/blobs/...?X-Amz-Expires=900&...
        upload_expires_at:
          type: string
          format: date-time
    FileDownloadResponse:
      type: object
      properties:
        id:
          type: string
          example: file_xyz789
        environment_id:
          type: string
          example: env_def456
        path:
          type: string
          example: config/example.txt
        content_hash:
          type: string
          description: Expected SHA-256 hash to verify after download (and post-M4 decryption)
          example: a1b2c3d4e5f6...
        size:
          type: integer
          example: 4096
        pushed_by:
          type: string
          example: usr_def456
        pushed_at:
          type: string
          format: date-time
        download_url:
          type: string
          description: Short-lived signed GET URL (15 min expiry)
          example: https://r2.invosit.dev/blobs/a1b2c3?X-Amz-Expires=900&...
        download_expires_at:
          type: string
          format: date-time
...